+++ app/src/main/java/com/ikemura23/yaeyama/linerchecker/ui/compose/YaimafuniApp.kt
@@ -1,12 +1,8 @@
 package com.ikemura23.yaeyama.linerchecker.ui.compose
 
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.ikemura23.yaeyama.linerchecker.ui.compose.dashboard.DashboardScreen
-import com.ikemura23.yaeyama.linerchecker.ui.compose.portstatusdetail.LinerStatusDetailScreen
+import com.ikemura23.yaeyama.linerchecker.ui.compose.navigation.YaimafuniNavHost
 
 @Composable
 fun YaimafuniApp() {
@@ -14,29 +10,4 @@ fun YaimafuniApp() {
     YaimafuniNavHost(
         navController = navController
     )
-}
-
-@Composable
-fun YaimafuniNavHost(
-    navController: NavHostController
-) {
-    NavHost(navController = navController, startDestination = Screen.Dashboard.route) {
-        composable(route = Screen.Dashboard.route) {
-            DashboardScreen(
-                onRowClick = {
-                    // TODO: routeId を引数にわたす
-                    navController.navigate(
-                        Screen.PortStatusDetail.createRoute("tempRouteId")
-                    )
-                }
-            )
-        }
-
-        composable(
-            route = Screen.PortStatusDetail.route,
-            arguments = Screen.PortStatusDetail.navArguments,
-        ) {
-            LinerStatusDetailScreen()
-        }
-    }
 }
